# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Couple Budget Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Accounts ──────────────────────────────────────────────────────────────
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Transactions ──────────────────────────────────────────────────────────
...
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: List budgets, optionally filtered by month
      parameters:
        - name: month
          in: query
          schema: { type: ["string", "null"] }
          description: "YYYY-MM format"
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create or upsert a budget entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Bills ─────────────────────────────────────────────────────────────────
  /bills:
    get:
      operationId: getBills
      tags: [bills]
      summary: List all bills and subscriptions
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a bill or subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillUpdate"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /bills/{id}/mark-paid:
    post:
      operationId: markBillPaid
      tags: [bills]
      summary: Mark a bill as paid for the current cycle
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"

  # ── Savings Goals ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, color, icon, expenseType]
      properties:
        id:          { type: integer }
        name:        { type: string }
        color:       { type: string }
        icon:        { type: string }
        expenseType: { type: string, enum: [fixed, variable, wants, income] }
        sortOrder:   { type: integer }
    CategoryInput:
      type: object
      required: [name, color, icon, expenseType]
      properties:
        name:        { type: string }
        color:       { type: string }
        icon:        { type: string }
        expenseType: { type: string, enum: [fixed, variable, wants, income] }
        sortOrder:   { type: ["integer", "null"] }
    CategoryUpdate:
      type: object
      properties:
        name:        { type: string }
        color:       { type: string }
        icon:        { type: string }
        expenseType: { type: string, enum: [fixed, variable, wants, income] }
        sortOrder:   { type: ["integer", "null"] }

    # ── Accounts ──────────────────────────────────────────────────────────────
    Account:
      type: object
      required: [id, name, type]
      properties:
        id:       { type: integer }
        name:     { type: string }
        type:     { type: string, enum: [checking, savings, credit, joint] }
        ownerId:  { type: ["integer", "null"] }
        balance:  { type: ["number", "null"] }
        isJoint:  { type: boolean }
    AccountInput:
      type: object
      required: [name, type]
      properties:
        name:     { type: string }
        type:     { type: string, enum: [checking, savings, credit, joint] }
        ownerId:  { type: ["integer", "null"] }
        balance:  { type: ["number", "null"] }
        isJoint:  { type: boolean }
    AccountUpdate:
      type: object
      properties:
        name:     { type: string }
        type:     { type: string, enum: [checking, savings, credit, joint] }
        ownerId:  { type: ["integer", "null"] }
        balance:  { type: ["number", "null"] }
        isJoint:  { type: boolean }

    # ── Transactions ──────────────────────────────────────────────────────────
...
    Budget:
      type: object
      required: [id, month, categoryId, amount]
      properties:
        id:           { type: integer }
        month:        { type: string }
        categoryId:   { type: integer }
        amount:       { type: number }
        categoryName: { type: ["string", "null"] }
        categoryColor: { type: ["string", "null"] }
        rolloverEnabled: { type: boolean }
        rolloverAmount:  { type: ["number", "null"] }
    BudgetInput:
      type: object
      required: [month, categoryId, amount]
      properties:
        month:           { type: string }
        categoryId:      { type: integer }
        amount:          { type: number }
        rolloverEnabled: { type: boolean }
    BudgetUpdate:
      type: object
      properties:
        amount:          { type: number }
        rolloverEnabled: { type: boolean }

    # ── Bills ─────────────────────────────────────────────────────────────────
    Bill:
      type: object
      required: [id, name, amount, dueDay, frequency, ownership, paidById]
      properties:
        id:            { type: integer }
        name:          { type: string }
        amount:        { type: number }
        dueDay:        { type: integer }
        frequency:     { type: string, enum: [monthly, quarterly, annual, weekly] }
        ownership:     { type: string, enum: [shared, personal] }
        paidById:      { type: integer }
        categoryId:    { type: ["integer", "null"] }
        isActive:      { type: boolean }
        isPaidThisCycle: { type: boolean }
        lastPaidDate:  { type: ["string", "null"] }
        notes:         { type: ["string", "null"] }
        categoryName:  { type: ["string", "null"] }
        paidByName:    { type: ["string", "null"] }
        nextDueDate:   { type: ["string", "null"] }
    BillInput:
      type: object
      required: [name, amount, dueDay, frequency, ownership, paidById]
      properties:
        name:       { type: string }
        amount:     { type: number }
        dueDay:     { type: integer }
        frequency:  { type: string, enum: [monthly, quarterly, annual, weekly] }
        ownership:  { type: string, enum: [shared, personal] }
        paidById:   { type: integer }
        categoryId: { type: ["integer", "null"] }
        isActive:   { type: boolean }
        notes:      { type: ["string", "null"] }
    BillUpdate:
      type: object
      properties:
        name:       { type: string }
        amount:     { type: number }
        dueDay:     { type: integer }
        frequency:  { type: string, enum: [monthly, quarterly, annual, weekly] }
        ownership:  { type: string, enum: [shared, personal] }
        paidById:   { type: integer }
        categoryId: { type: ["integer", "null"] }
        isActive:   { type: boolean }
        notes:      { type: ["string", "null"] }

    # ── Savings Goals ─────────────────────────────────────────────────────────
...