# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Couple Budget Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputType } from "./accountInputType";

export interface AccountInput {
  name: string;
  type: AccountInputType;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  balance?: number | null;
  isJoint?: boolean;
}